Move util to test_utils
