+++ store/src/commonTest/kotlin/org/mobilenativefoundation/store/store5/ClearStoreByKeyTests.kt
@@ -5,9 +5,9 @@ import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import org.mobilenativefoundation.store.store5.StoreReadResponse.Data
-import org.mobilenativefoundation.store.store5.util.InMemoryPersister
-import org.mobilenativefoundation.store.store5.util.asSourceOfTruth
-import org.mobilenativefoundation.store.store5.util.getData
+import org.mobilenativefoundation.store.store5.test_utils.InMemoryPersister
+import org.mobilenativefoundation.store.store5.test_utils.asSourceOfTruth
+import org.mobilenativefoundation.store.store5.test_utils.getData
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
