+++ store/src/commonTest/kotlin/org/mobilenativefoundation/store/store5/test_utils/TestStoreExt.kt
@@ -1,4 +1,4 @@
-package org.mobilenativefoundation.store.store5.util
+package org.mobilenativefoundation.store.store5.test_utils
 
 import kotlinx.coroutines.flow.filterNot
 import kotlinx.coroutines.flow.first
