+++ store/src/commonTest/kotlin/org/mobilenativefoundation/store/store5/test_utils/fake/NotesDatabase.kt
@@ -1,7 +1,7 @@
-package org.mobilenativefoundation.store.store5.util.fake
+package org.mobilenativefoundation.store.store5.test_utils.fake
 
-import org.mobilenativefoundation.store.store5.util.model.InputNote
-import org.mobilenativefoundation.store.store5.util.model.OutputNote
+import org.mobilenativefoundation.store.store5.test_utils.model.InputNote
+import org.mobilenativefoundation.store.store5.test_utils.model.OutputNote
 
 internal class NotesDatabase {
     private val db: MutableMap<NotesKey, OutputNote?> = mutableMapOf()
