+++ store/src/commonTest/kotlin/org/mobilenativefoundation/store/store5/test_utils/fake/NotesUpdaterProvider.kt
@@ -1,10 +1,10 @@
-package org.mobilenativefoundation.store.store5.util.fake
+package org.mobilenativefoundation.store.store5.test_utils.fake
 
 import org.mobilenativefoundation.store.store5.Updater
 import org.mobilenativefoundation.store.store5.UpdaterResult
-import org.mobilenativefoundation.store.store5.util.model.InputNote
-import org.mobilenativefoundation.store.store5.util.model.NotesWriteResponse
-import org.mobilenativefoundation.store.store5.util.model.OutputNote
+import org.mobilenativefoundation.store.store5.test_utils.model.InputNote
+import org.mobilenativefoundation.store.store5.test_utils.model.NotesWriteResponse
+import org.mobilenativefoundation.store.store5.test_utils.model.OutputNote
 
 internal class NotesUpdaterProvider(private val api: NotesApi) {
     fun provide(): Updater<NotesKey, OutputNote, NotesWriteResponse> =
