+++ store/src/commonTest/kotlin/org/mobilenativefoundation/store/store5/test_utils/fake/fallback/SecondaryPagesApi.kt
@@ -1,4 +1,4 @@
-package org.mobilenativefoundation.store.store5.util.fake.fallback
+package org.mobilenativefoundation.store.store5.test_utils.fake.fallback
 
 class SecondaryPagesApi() {
     val name: String = "SecondaryPagesApi"
